fix: move custom providers to LayoutProvider wrapper
